Source push - DU About
Based on DU Certified, this allows for us to update our device and personnel roster
in real time instead of having to update the app every time someone parts way with a
device or we part ways with a maintainer/developer.

Nothing to really update or fix on our end. Reason for pushing is so others can base
off it and expand on it.